pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dftlearn"]

[project]
name = "dft-learn"
version = "0.2.0"
description = "Machine learning and data analysis for DFT calculations with GPU acceleration and sklearn compatibility"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
maintainers = [
    {name = "Harlan Heilman", email="harlan.heilman@wsu.edu"}
]
keywords = ["dft", "nexafs", "spectroscopy", "machine-learning", "gpu", "sklearn"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
]

# Core dependencies - REQUIRED
dependencies = [
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.5.0,<1.8.0",
    "pandas>=2.1.0",
    "numba>=0.61.2",
    "matplotlib>=3.10.0",
    "periodictable>=1.6.0",
]

[project.optional-dependencies]
# GPU support - OPTIONAL but recommended for large datasets
gpu = [
    "cupy-cuda12x>=13.6.0",
]

# Deep learning features (for future extensions)
torch = [
    "torch>=2.8.0",
    "torchvision>=0.19.0",
]

# Full scientific stack
full = [
    "polars>=0.20.0",
    "xarray>=2024.1.0",
    "h5py>=3.10.0",
    "h5netcdf>=1.6.0",
    "pyarrow>=14.0.0",
]

# Visualization extras
viz = [
    "py3dmol>=2.4.2",
    "seaborn>=0.13.0",
]

# Complete package with all features
all = [
    "dft-learn[gpu,torch,full,viz]",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true


[tool.ruff]
exclude = [
  ".github",
  ".git",
  ".vscode",
  ".venv",
  "docs",
  "scripts",
  "tests",
  "python_pipeline",
]
fix = true
line-length = 88

[tool.pyrefly]
exclude = [
  ".github",
  ".git",
  ".vscode",
  ".venv",
  "docs",
  "scripts",
  "tests",
  "python_pipeline",
]

[tool.ruff.lint]
ignore = ["D200"]
select = [
  "E", # pycodestyle
  "W", # pycodestyle
  "F", # Pyflakes
  "B", # flake8-bugbear
  "C4", # flake8-comprehensions
  "D", # flake8-docstrings
  "I", # isort
  "SIM", # flake8-simplify
  "TCH", # flake8-type-checking
  "TID", # flake8-tidy-imports
  "UP", # pyupgrade
  "PT", # flake8-pytest-style
  "RUF", # Ruff-specific rules
  "PTH", # flake8-use-pathlib
  "FA", # flake8-future-annotations
  "PIE", # flake8-pie
  "TD", # flake8-todos
  "TRY", # tryceratops
  "EM", # flake8-errmsg
  "FBT001", # flake8-boolean-trap
]
[tool.ruff.lint.pycodestyle]
max-doc-length = 88

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.flake8-type-checking]
strict = true

[tool.ruff.format]
docstring-code-format = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "--maxfail=1",
    "--disable-warnings",
    "-q",
    "--strict-markers",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests requiring GPU (deselect with '-m \"not gpu\"')",
    "integration: marks integration tests",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-order>=1.3.0",
    "pytest-cov>=4.1.0",
    "pytest-benchmark>=4.0.0",
    "ruff>=0.8.0",
    "mypy>=1.8.0",
    "ipython>=8.20.0",
    "jupyterlab>=4.4.0",
]
